refactor(crf): switch all id fields from i32 to i64 to match BIGSERIAL
The crf_* migrations declare every primary key as BIGSERIAL and every
foreign key as BIGINT, but the Rust types were i32. Update the domain,
usecase, apis, and aegis-server layers to match the storage shape.

Scope: only fields that identify a row move to i64. Non-id integers
(`CrfForm.order`, `CrfItem.order`) stay i32.

Affected types:
  - aggregates: id / version_id / form_id / item_id / option_id /
    unit_id / domain_annotation_id
  - domain ports: every id-bearing method parameter
  - apis DTOs / views / AnnotationOwner / CrfApiError variants / trait
    method signatures (including the seven `delete_*(id: i64)` methods)
  - usecase commands / views / method parameters
  - Postgres Row structs and SQLx binds
  - utoipa path parameter type annotations

Verification:
  cargo fmt --check  (all three crates clean)
  cargo clippy -p crf/apis/aegis-server --all-targets -- -D warnings
  cargo test -p crf --lib                      (66 passed)
  cargo test -p crf --test public_api          (8 passed)
  cargo test -p aegis-server --lib crf         (15 CRF tests passed)

The 28 pre-existing failures in dto roundtrip / auth / openapi tests
predate this refactor (verified by `git stash` + re-run); they are
serde / camelCase issues unrelated to the i32 → i64 change.

+++ apps/server/aegis-server/src/state.rs
@@ -297,7 +297,7 @@ pub(crate) mod test_support {
         ) -> Result<apis::crf::CrfVersionView, apis::crf::CrfApiError> {
             unimplemented!()
         }
-        async fn delete_version(&self, _id: i32) -> Result<(), apis::crf::CrfApiError> {
+        async fn delete_version(&self, _id: i64) -> Result<(), apis::crf::CrfApiError> {
             unimplemented!()
         }
         async fn create_form(
@@ -324,7 +324,7 @@ pub(crate) mod test_support {
         ) -> Result<apis::crf::CrfFormView, apis::crf::CrfApiError> {
             unimplemented!()
         }
-        async fn delete_form(&self, _id: i32) -> Result<(), apis::crf::CrfApiError> {
+        async fn delete_form(&self, _id: i64) -> Result<(), apis::crf::CrfApiError> {
             unimplemented!()
         }
         async fn create_item(
@@ -351,7 +351,7 @@ pub(crate) mod test_support {
         ) -> Result<apis::crf::CrfItemView, apis::crf::CrfApiError> {
             unimplemented!()
         }
-        async fn delete_item(&self, _id: i32) -> Result<(), apis::crf::CrfApiError> {
+        async fn delete_item(&self, _id: i64) -> Result<(), apis::crf::CrfApiError> {
             unimplemented!()
         }
         async fn create_option(
@@ -378,7 +378,7 @@ pub(crate) mod test_support {
         ) -> Result<apis::crf::CrfOptionView, apis::crf::CrfApiError> {
             unimplemented!()
         }
-        async fn delete_option(&self, _id: i32) -> Result<(), apis::crf::CrfApiError> {
+        async fn delete_option(&self, _id: i64) -> Result<(), apis::crf::CrfApiError> {
             unimplemented!()
         }
         async fn create_unit(
@@ -405,7 +405,7 @@ pub(crate) mod test_support {
         ) -> Result<apis::crf::CrfUnitView, apis::crf::CrfApiError> {
             unimplemented!()
         }
-        async fn delete_unit(&self, _id: i32) -> Result<(), apis::crf::CrfApiError> {
+        async fn delete_unit(&self, _id: i64) -> Result<(), apis::crf::CrfApiError> {
             unimplemented!()
         }
         async fn create_domain_annotation(
@@ -432,7 +432,7 @@ pub(crate) mod test_support {
         ) -> Result<apis::crf::DomainAnnotationView, apis::crf::CrfApiError> {
             unimplemented!()
         }
-        async fn delete_domain_annotation(&self, _id: i32) -> Result<(), apis::crf::CrfApiError> {
+        async fn delete_domain_annotation(&self, _id: i64) -> Result<(), apis::crf::CrfApiError> {
             unimplemented!()
         }
         async fn create_annotation(
@@ -477,7 +477,7 @@ pub(crate) mod test_support {
         ) -> Result<apis::crf::AnnotationView, apis::crf::CrfApiError> {
             unimplemented!()
         }
-        async fn delete_annotation(&self, _id: i32) -> Result<(), apis::crf::CrfApiError> {
+        async fn delete_annotation(&self, _id: i64) -> Result<(), apis::crf::CrfApiError> {
             unimplemented!()
         }
         async fn search_forms_by_version(
